requirements.txt
fastapi
uvicorn
python-multipart
requests
beautifulsoup4
Pillow
numpy
moviepy
imageio-ffmpeg
pymupdf
sqlalchemy
psycopg2-binary
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
